LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-align-stretching-replaced-items-expected.txt

This test checks that the alignment properties align-self and justify-self apply the 'stretch' value correctly on replaced elements.

The blue image's original size is 100px x 100px, default alignment is resolved as 'start' for replaced elements so it prevents stretching to be applied.


The blue image's original size is 100px x 100px, 'normal' is resolved as 'start' for replaced elements so it prevents stretching to be applied.


The blue image's original size is 100px x 100px, but it should be stretched to fill the 500px x 500px grid area it's placed into.


The blue image's original size is 100px x 100px, non-stretch values prevent stretching to be applied.


The blue image's original size is 100px x 100px, non-auto sizes prevent stretching to be applied.


The blue image's original size is 100px x 100px, auto-margins prevent stretching to be applied.



PASS .grid 1
PASS .grid 2
PASS .grid 3
PASS .grid 4
PASS .grid 5
PASS .grid 6
